Move common routines into a separate file
